pyproject.toml
[tool.poetry]
name = "netatmo2mqtt-app"
version = "0.1.0"
description = ""
authors = ["Dusan Hlavaty <dhlavaty@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
paho-mqtt = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
